+++ NNControl/Network/NeuralNetworkView.cs
@@ -12,7 +12,7 @@ namespace NNControl.Network
         public List<LayerView> Layers = new List<LayerView>();
         public List<NeuronView> SelectedNeuron = new List<NeuronView>();
         public List<SynapseView> HighlightedSynapses = new List<SynapseView>();
-        public virtual double Zoom { get; internal set; } = 0;
+        public virtual float Zoom { get; internal set; } = 0;
         public int NeuronsCount;
         public SynapseView SelectedSynapse;
